Tweaks and comments
This may point out a significant bug that arose as a consequence of Matlab->Julia translation in handling of the exponential of a matrix: Julia's `exp(A)` corresponds to Matlab's `expm(A)` (note the `m`).

There are also some inference issues that may be hurting performance. You can find them by profiling the first call to `zernike_img_fit` in your tests.